+++ tests/testthat/test-user.R
@@ -2,9 +2,8 @@ context("User stuff")
 
 with_mock_HTTP({
     test_that("Getting user object", {
-        user <- getUser("/api/users/user1.json")
-        expect_true(inherits(user, "ShojiObject"))
-        expect_identical(user@body$email, "fake.user@example.com")
+        expect_true(inherits(me(), "UserEntity"))
+        expect_identical(email(me()), "fake.user@example.com")
     })
 
     test_that("Getting account's user catalog", {
@@ -27,43 +26,21 @@ with_mock_HTTP({
 if (run.integration.tests) {
     with(test.authentication, {
         test_that("User can be fetched", {
-            user <- try(getUser())
-            expect_true(inherits(user, "ShojiObject"))
+            expect_true(inherits(me(), "UserEntity"))
         })
 
-        u.email <- uniqueEmail()
-        u.name <- now()
-        u.url <- try(invite(u.email, name=u.name, notify=FALSE))
-
-        test_that("User can be invited", {
-            skip_on_jenkins("Jenkins user needs more permissions")
-            usercat <- getAccountUserCatalog()
-            expect_true(u.url %in% urls(usercat))
-            expect_true(u.email %in% emails(usercat))
-            expect_true(u.name %in% sub(" +$", "", names(usercat)))
-        })
-
-        test_that("User can be deleted", {
-            skip_on_jenkins("Jenkins user needs more permissions")
-            try(crDELETE(u.url))
-            usercat <- refresh(getAccountUserCatalog())
-            expect_false(u.url %in% urls(usercat))
-            expect_false(u.email %in% emails(usercat))
-            expect_false(u.name %in% sub(" +$", "", names(usercat)))
-        })
-
-        test_that("test.user() setup/teardown", {
+        test_that("Create and delete user; cleanup(testUser()) setup/teardown", {
             skip_on_jenkins("Jenkins user needs more permissions")
             u.email <- paste0("test+", as.numeric(Sys.time()), "@crunch.io")
             u.name <- now()
             usercat <- getAccountUserCatalog()
             expect_false(u.email %in% emails(usercat))
             expect_false(u.name %in% sub(" +$", "", names(usercat)))
-            with(test.user(u.email, u.name), {
+            with(cleanup(testUser(u.email, u.name)), as="u", {
                 usercat <- refresh(usercat)
                 expect_true(u.email %in% emails(usercat))
                 expect_true(u.name %in% sub(" +$", "", names(usercat)))
-                user <- index(usercat)[[u]]
+                user <- index(usercat)[[self(u)]]
                 expect_false(user$account_permissions$create_datasets)
                 expect_false(user$account_permissions$alter_users)
             })
@@ -74,18 +51,18 @@ if (run.integration.tests) {
 
         test_that("User with permissions", {
             skip_on_jenkins("Jenkins user needs more permissions")
-            with(test.user(advanced=TRUE), {
-                user <- index(getAccountUserCatalog())[[u]]
+            with(cleanup(testUser(advanced=TRUE)), as="u", {
+                user <- index(getAccountUserCatalog())[[self(u)]]
                 expect_true(user$account_permissions$create_datasets)
                 expect_false(user$account_permissions$alter_users)
             })
-            with(test.user(admin=TRUE), {
-                user <- index(getAccountUserCatalog())[[u]]
+            with(cleanup(testUser(admin=TRUE)), as="u", {
+                user <- index(getAccountUserCatalog())[[self(u)]]
                 expect_false(user$account_permissions$create_datasets)
                 expect_true(user$account_permissions$alter_users)
             })
-            with(test.user(admin=TRUE, advanced=TRUE), {
-                user <- index(getAccountUserCatalog())[[u]]
+            with(cleanup(testUser(admin=TRUE, advanced=TRUE)), as="u", {
+                user <- index(getAccountUserCatalog())[[self(u)]]
                 expect_true(user$account_permissions$create_datasets)
                 expect_true(user$account_permissions$alter_users)
             })
@@ -94,7 +71,7 @@ if (run.integration.tests) {
 
     test_that("User cannot be fetched if logged out", {
         logout()
-        expect_error(getUser(),
+        expect_error(me(),
             "You must authenticate before making this request")
     })
 }