+++ NEWS.md
@@ -1,5 +1,11 @@
 ### crunch 1.9.13 (under development)
 * Add `owner` and `owner<-` for datasets to read and modify the owner
+* Add `owners` and `ownerNames` for DatasetCatalog
+* `is.editor` and `is.editor<-` for project MemberCatalog
+* `me` function to get the user entity for yourself
+* Add missing print method for DatasetOrder
+* Support creating OrderGroups (for both Datasets and Variables) by assigning URLs into a new group name
+* Fix bug in setting nested groups inside DatasetOrder
 * Fix failure on interactive login in R.app on OS X
 
 ### crunch 1.9.12